schedule_core/requirements.txt
aiohttp==3.9.1
aioredis==1.3.1
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.8
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.4
billiard==4.1.0
bleach==6.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
combu==1.2.1
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.4
cssselect==1.2.0
daphne==3.0.2
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.7
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-decouple==2.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
drf-yasg==1.21.7
flake8==6.1.0
flake8-django==1.4
flower==2.0.1
frozenlist==1.4.1
future==0.18.3
geocoder==1.38.1
geoip2==4.8.0
h11==0.14.0
h2==4.1.0
hiredis==2.3.2
hpack==4.0.0
humanize==4.8.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==7.0.1
incremental==22.10.0
inflection==0.5.1
ip2geotools==0.1.6
IP2Location==8.10.1
isort==5.13.2
jaraco.classes==3.3.0
keyring==24.3.0
kombu==5.3.2
lazy-object-proxy==1.10.0
lxml==4.9.4
markdown-it-py==3.0.0
maxminddb==2.5.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
nh3==0.2.15
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
pip-review==1.3.0
pkginfo==1.9.6
platformdirs==4.1.0
priority==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==1.7.1
pylint==3.0.3
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyquery==2.0.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
ratelim==0.1.6
readme-renderer==42.0
redis==4.6.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
selenium==4.16.0
sentry-sdk==1.32.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tomlkit==0.12.3
tornado==6.3.3
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
twine==4.0.2
Twisted==23.10.0
txaio==23.1.1
typed-ast==1.5.5
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
webencodings==0.5.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
